refactor: PotController에서 Repository 호출 제거
