+++ src/main/java/stackpot/stackpot/pot/controller/PotMemberController.java
@@ -7,11 +7,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import stackpot.stackpot.apiPayload.ApiResponse;
-import stackpot.stackpot.pot.service.PotMemberService;
+import stackpot.stackpot.pot.service.PotMemberCommandService;
 import stackpot.stackpot.pot.dto.PotMemberAppealResponseDto;
 import stackpot.stackpot.pot.dto.PotMemberInfoResponseDto;
 import stackpot.stackpot.pot.dto.PotMemberRequestDto;
 import stackpot.stackpot.pot.dto.UpdateAppealRequestDto;
+import stackpot.stackpot.pot.service.PotMemberQueryService;
 
 import java.util.List;
 
@@ -21,15 +22,16 @@ import java.util.List;
 @Tag(name = "Pot Member Management", description = "팟 멤버 관리 API")
 public class PotMemberController {
 
-    private final PotMemberService potMemberService;
+    private final PotMemberCommandService potMemberCommandService;
+    private final PotMemberQueryService potMemberQueryService;
 
     @Operation(summary = "팟 멤버 정보 (KAKAOID, 닉네임) 조회 API")
-    @GetMapping // ✅ @PathVariable을 사용하려면 URL에 포함해야 함
+    @GetMapping
     public ResponseEntity<ApiResponse<List<PotMemberInfoResponseDto>>> getPotMembers(
             @PathVariable("pot_id") Long potId) {
 
-        //potMemberService.validateIsOwner(potId); // 팟 생성자 검증 추가
-        List<PotMemberInfoResponseDto> response = potMemberService.getPotMembers(potId);
+
+        List<PotMemberInfoResponseDto> response = potMemberQueryService.getPotMembers(potId);
 
         return ResponseEntity.ok(ApiResponse.onSuccess(response));
     }
@@ -44,7 +46,7 @@ public class PotMemberController {
     public ResponseEntity<ApiResponse<List<PotMemberAppealResponseDto>>> addPotMembers(
             @PathVariable("pot_id") Long potId,
             @RequestBody @Valid PotMemberRequestDto requestDto) {
-        List<PotMemberAppealResponseDto> response = potMemberService.addMembersToPot(potId, requestDto);
+        List<PotMemberAppealResponseDto> response = potMemberCommandService.addMembersToPot(potId, requestDto);
         return ResponseEntity.ok(ApiResponse.onSuccess(response));
     }
     @Operation(summary = "팟 어필 API")
@@ -53,7 +55,7 @@ public class PotMemberController {
             @PathVariable("pot_id") Long potId,
 
             @RequestBody @Valid UpdateAppealRequestDto requestDto) {
-        potMemberService.updateAppealContent(potId, requestDto.getAppealContent());
+        potMemberCommandService.updateAppealContent(potId, requestDto.getAppealContent());
         return ResponseEntity.ok(ApiResponse.onSuccess("어필 내용이 성공적으로 업데이트되었습니다."));
     }
 }
