+++ src/main/java/stackpot/stackpot/pot/service/PotMemberQueryService.java
@@ -1,0 +1,9 @@
+package stackpot.stackpot.pot.service;
+
+import stackpot.stackpot.pot.dto.PotMemberInfoResponseDto;
+
+import java.util.List;
+
+public interface PotMemberQueryService {
+    List<PotMemberInfoResponseDto> getPotMembers(Long potId);
+}
